% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customswitchButton.R
\name{customswitchButton}
\alias{customswitchButton}
\title{Customised TRUE-FALSE switch button for Rshiny
Only sing CSS3 code (No javascript)}
\usage{
customswitchButton(inputId, label, value = FALSE, col = "GB", type = "TF")
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value (TRUE or FALSE).}

\item{col}{Color set of the switch button. Choose between "GB" (Grey-Blue) and "RG" (Red-Green)}

\item{type}{Text type of the button. Choose between "TF" (TRUE - FALSE), "OO" (ON - OFF) or leave empty for no text.}
}
\description{
Sébastien Rochette
http://statnmap.com/en/
April 2016
}
\details{
CSS3 code was found on https://proto.io/freebies/onoff/
For CSS3 customisation, refer to this website.

A function to change the Original checkbox of rshiny
into a nice true/false or on/off switch button
No javascript involved. Only CSS code.

To be used with CSS script 'button.css' stored in a 'www' folder in your Shiny app folder
}
