% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawChromatogram.R
\name{drawChromatogram.DrawAlignR}
\alias{drawChromatogram.DrawAlignR}
\title{drawChromatogram method for class DrawAlignR}
\usage{
\method{drawChromatogram}{DrawAlignR}(app.obj, type = "default")
}
\arguments{
\item{app.obj}{(ReactiveValues/list) Object containing app related inputs and parameters, of class DrawAlignR. TODO Add list of inputs with details.}
}
\description{
drawChromatogram method for class DrawAlignR
}
