% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawChromatogram.R
\name{drawChromatogram}
\alias{drawChromatogram}
\title{Generic drawChromatogram definition}
\usage{
drawChromatogram(app.obj, ...)
}
\arguments{
\item{app.obj}{}
}
\description{
Generic drawChromatogram definition
}
