% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChromatogramDataPoints.R
\name{getChromatogramDataPoints_}
\alias{getChromatogramDataPoints_}
\title{Extract Chromatographic data (Retention Time and Intensity) from mzML or sqMass chromatograms.
This function can be used to extract chromatogram retention and intensity values for a given
list of fragment ids.}
\usage{
getChromatogramDataPoints_(filename, frag_ids)
}
\arguments{
\item{filename}{A character vector of the absolute path and filename of the chromatogram file. (Must be .mzML or sqMass format)}

\item{frag_ids}{A list of a vector containing fragment ids}
}
\value{
A list of fragment ids containing 2 vectors for Retention time and Intensity
}
\description{
Extract Chromatographic data (Retention Time and Intensity) from mzML or sqMass chromatograms.
This function can be used to extract chromatogram retention and intensity values for a given
list of fragment ids.
}
\author{
Justin Sing \url{https://github.com/singjc}
}
