% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getFunctionCallArgs}
\alias{getFunctionCallArgs}
\title{Get string of function call}
\usage{
getFunctionCallArgs(func_args_list)
}
\arguments{
\item{func_args_list}{A list object containing information of arguments to a function call}
}
\value{
A chracter vector representin function call
}
\description{
Get string of function call
}
