% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getMaintainer}
\alias{getMaintainer}
\title{Get string information for Maintainer}
\usage{
getMaintainer(
  email = "justincsing@gmail.com",
  github = "https://github.com/singjc/DrawAlignR"
)
}
\arguments{
\item{email}{A character vector for email address to contact maintainer}

\item{github}{A character vector for github repository to submit issue/feature requests}
}
\value{
A chracter vector with Maintainer information
}
\description{
Get string information for Maintainer
}
