% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPepLibData.R
\name{getPepLibData_}
\alias{getPepLibData_}
\title{Extract data from a PQP library file
This function can be used to extract information from the PQP library file}
\usage{
getPepLibData_(libfile, peptide_id = "")
}
\arguments{
\item{libfile}{A character vector of the absolute path and filename of the library file. (Must be .pqp format)}

\item{peptide_id}{A character vector for extraction of a specific peptide. I.e. 'ANSSPTTNIDHLK'}
}
\value{
A data.table containing spectral library information
}
\description{
Extract data from a PQP library file
This function can be used to extract information from the PQP library file
}
\author{
Justin Sing \url{https://github.com/singjc}
}
