% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXIC.r
\name{getXIC}
\alias{getXIC}
\title{Plot and Extracted Ion Chromatogram
This function can be used to plot an Extracted Ion Chromatogram}
\usage{
getXIC(
  graphic_obj,
  mod,
  df_lib,
  chromatogram_file,
  in_osw = NULL,
  transition_type = "detecting",
  intersecting_mz = NULL,
  uni_mod_list = NULL,
  max_Int = NULL,
  smooth_chromatogram = TRUE,
  doFacetZoom = FALSE,
  FacetFcnCall = NULL,
  top_trans_mod_list = NULL,
  Isoform_Target_Charge = NULL,
  RT_pkgrps = NULL,
  plotIdentifying.Unique = NULL,
  plotIdentifying.Shared = NULL,
  plotIdentifying.Against = NULL,
  show_n_transitions = NULL,
  show_legend = T,
  verbosity = 0
)
}
\arguments{
\item{graphic_obj}{A ggplot() graphics handle. Initialize with g <- ggplot() to create an empty ggplot handle}

\item{mod}{A character vector for specific peptide/modified peptide to extract I.e. 'ANSSPTTNIDHLK'/'ANS(UniMod:21)SPTTNIDHLK(UniMod:259)'. The MODIFIED_SEQUENCE column is used}

\item{df_lib}{A data.table containing spectral library information}

\item{chromatogram_file}{A character vector of the absolute path and filename of the chromatogram file. (Must be .mzML or sqMass format)}

\item{in_osw}{A character vector of the absolute path and filename of the OpenSwath Output file. (Must be .osw) @TODO maybe make this more robust for tsv files as well?}

\item{@TODO}{will add more decriptions for other params soon}

\item{df}{A data.frame/data.table/matrix object to check for number of rows}

\item{msg}{error message to return}

\item{numeric_obj}{Check if an numeric object is not NULL a}
}
\value{
A list containing graphic_obj = the graphic handle for the ggplot filled with data and max_Int = the maximun intensity

if data.frame has 0 rows, return true, otherwise return false

If numeric object is not null, round numeric object to 4 digits
}
\description{
Plot and Extracted Ion Chromatogram
This function can be used to plot an Extracted Ion Chromatogram
}
\author{
Justin Sing \url{https://github.com/singjc}

Check Dataframe, if empty return an object and stop function call
}
