% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aligned.R
\name{plot_aligned}
\alias{plot_aligned}
\title{A function for the visualization of an aligned chromatogram to a reference chromatogram}
\usage{
plot_aligned(chrom, shift, precursor, Run_ID, Reference_ID, RT,
  Reference_RT, Left_width, Right_width, mz, sequence)
}
\arguments{
\item{chrom}{rds chromatogram file which contains a list of peptides with their retention times and intensities.}

\item{shift}{The specific retention time shift of the alignment, obtained upstream using DialignR}

\item{RT}{The retention time of the peptide}

\item{Reference_RT}{The retention time of the precursor in the reference run.}

\item{Left_width}{Left width of the retention time}

\item{Right_width}{Right width of the retention time}

\item{mz}{The mass to charge value of the precursor}

\item{peptide}{the precursor id of the specific peptide you would like to plot.}

\item{Run_Id}{The run from which the peptide came from}

\item{Reference_Id}{The reference run/precursor to which this peptide was aligned to.}

\item{The}{sequence of the precursor}
}
\value{
None. Produces plot output.
}
\description{
A function for the visualization of an aligned chromatogram to a reference chromatogram
}
