% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrom_reference.R
\name{plot_chrom_reference}
\alias{plot_chrom_reference}
\title{A function for plotting the ms2 chromatogram of any peptide. It's usage is to plot the reference chromatogram,
which is the run that all other chromatograms are aligned to.}
\usage{
plot_chrom_reference(chrom, precursor, Run_ID, RT, Left_width, Right_width,
  mz, sequence)
}
\arguments{
\item{chrom}{rds chromatogram file which contains a list of peptides with their retention times and intensities.}

\item{RT}{The retention time of the peptide}

\item{Left_width}{Left width of the retention time}

\item{Right_width}{Right width of the retention time}

\item{mz}{The mass to charge value of the precursor}

\item{peptide}{the precursor id of the specific peptide you would like to plot.}

\item{Run_Id}{The run from which the peptide came from}

\item{The}{sequence of the precursor}
}
\value{
None. Produces plot output.
}
\description{
A function for plotting the ms2 chromatogram of any peptide. It's usage is to plot the reference chromatogram,
which is the run that all other chromatograms are aligned to.
}
