% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqp_data_access.R
\docType{methods}
\name{pqp_data_access}
\alias{pqp_data_access}
\alias{sqmassdbConnection}
\alias{pqp_data_access-method}
\alias{getDataForAllChromatograms}
\title{Generic pqp Data Access Function}
\usage{
pqp_data_access(
  filename,
  init = NULL,
  obj = NULL,
  filter_decoy = 0,
  unmodified_sequence = "",
  modified_sequence = "",
  ...
)

\S4method{pqp_data_access}{character,logical,missing,missing,missing,missing}(filename, init)

\S4method{pqp_data_access}{character,missing,missing,ANY,ANY,ANY}(
  filename,
  filter_decoy = 0,
  unmodified_sequence = "",
  modified_sequence = ""
)
}
\arguments{
\item{filename}{(character) A character/string of the sqMass database file}

\item{init}{(logical) A logical value to establish a connection to the database, and return an S4 object with the connection}

\item{obj}{(S4) An S4 object that contains the filename, connection to the database, and the chromatoram data}

\item{filter_decoy}{(integer) An integer of 0 or 1. 0 will filter for only targets, 1 will filter for only decoys. Default: 0}

\item{unmodified_sequence}{(numeric) A character vector for extraction of a specific peptide. I.e. 'ANSSPTTNIDHLK'}

\item{modified_sequence}{(chracter) An array of two string vectors indicating a specific modified peptide sequence with both UniMod annotation and actual modification name to extract information for. I.e. c(ANS(Phos)SNSLK, ANS(UniMod:21)SNSLK) (Default: '')}
}
\description{
This function will dispatch one of the method-family functions based on input arguments. \cr
The main purpose of this function is to extract chromatographic data stored in the sqMass sql-based database. \cr
Get data from multiple chromatograms \cr
- compression is one of 0 = no, 1 = zlib, 2 = np-linear, 3 = np-slof, 4 = np-pic, 5 = np-linear + zlib, 6 = np-slof + zlib, 7 = np-pic + zlib \cr
- data_type is one of 0 = mz, 1 = int, 2 = rt \cr
- data contains the raw (blob) data for a single data array \cr

Get data from multiple chromatograms
}
