% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDrawAlignRShiny.R
\name{runDrawAlignR}
\alias{runDrawAlignR}
\title{Launches the shiny app for DrawAlignR - User facing function.}
\usage{
runDrawAlignR()
}
\value{
None. Runs the Shiny webapp found in inst/available-shiny-apps/DrawAlignR/app.R
}
\description{
Launches the shiny app for DrawAlignR - User facing function.
}
