% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqmass_data_access.R
\docType{methods}
\name{sqmass_data_access}
\alias{sqmass_data_access}
\alias{sqmassdbConnection}
\alias{sqmass_data_access-method}
\alias{returnDataForChromatogram}
\alias{getDataForAllChromatograms}
\alias{getDataForChromatogramsFromID}
\alias{getDataForChromatogramFromNativeId}
\title{Generic sqMass Data Access Function}
\usage{
sqmass_data_access(
  filename,
  init = NULL,
  obj = NULL,
  data = NULL,
  ids = NULL,
  native_id = NULL,
  ...
)

\S4method{sqmass_data_access}{character,logical,missing,missing,missing,missing}(filename, init)

\S4method{sqmass_data_access}{missing,missing,sqmass_data_access,data.frame,missing,missing}(obj, data)

\S4method{sqmass_data_access}{character,missing,missing,missing,missing,missing}(filename)

\S4method{sqmass_data_access}{character,missing,missing,missing,numeric,missing}(filename, ids)

\S4method{sqmass_data_access}{character,missing,missing,missing,missing,character}(filename, native_id)
}
\arguments{
\item{filename}{(character) A character/string of the sqMass database file}

\item{init}{(logical) A logical value to establish a connection to the database, and return an S4 object with the connection}

\item{obj}{(S4) An S4 object that contains the filename, connection to the database, and the chromatoram data}

\item{data}{(data.frame) A data.frame that contains the CHROMATOGRAM table information from the database}

\item{ids}{(numeric) A numeric vector of CHROMATOGRAM_ID(s) that you want to extract chromatogram data for}

\item{native_id}{(chracter) A chracter vector of NATIVE_ID(s) that you want to extract a chromatogram data for}
}
\description{
This function will dispatch one of the method-family functions based on input arguments. \cr
The main purpose of this function is to extract chromatographic data stored in the sqMass sql-based database. \cr
Get data from multiple chromatograms \cr
- compression is one of 0 = no, 1 = zlib, 2 = np-linear, 3 = np-slof, 4 = np-pic, 5 = np-linear + zlib, 6 = np-slof + zlib, 7 = np-pic + zlib \cr
- data_type is one of 0 = mz, 1 = int, 2 = rt \cr
- data contains the raw (blob) data for a single data array \cr

Get data from multiple chromatograms

Get data from multiple chromatograms or a single chromatogram given a list of chromatogram IDs

Get data from multiple chromatograms or a single chromatogram given a list of native IDs
}
