% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{text2numericInput}
\alias{text2numericInput}
\title{Convert Text Input separated by a comma to a numeric vector}
\usage{
text2numericInput(text)
}
\arguments{
\item{text}{A character vector containing numbers separated by a comma}
}
\value{
A numeric vector
}
\description{
Convert Text Input separated by a comma to a numeric vector
}
