% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withConsoleRedirect.R
\name{withConsoleRedirect}
\alias{withConsoleRedirect}
\title{Capture output from console}
\usage{
withConsoleRedirect(containerId, expr)
}
\arguments{
\item{containerId}{A character vector for UI container id to send text to.}

\item{expr}{An expression/function to wrap around that spits out messages}
}
\value{
A chracter vector with console output reversed
}
\description{
Capture output from console
}
